Moved the CardService-related methods and fields to a new CardService class. Confirmed tests still worked.
